ACTION REQUIRED
for the client-server stuff i moved a few things common to server and
client to a new shared project.

to tell eclipse where to find the things, add the shared project to the
build path of client and server. also make sure not to commit the
.classpath file etc. as this is a new project you also need to make the
initial configurations with the .gitignore file.